requirements.txt
alpaca-py==0.35.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
python-dotenv==1.0.1
PyYAML==6.0.2
pytz==2024.1
